test: step1 - 테스트 수정 및 기존 연관관계가 있는 findById 테스트 loader 테스트로 이관
