+++ src/test/java/persistence/entity/CustomJpaRepositoryTest.java
@@ -28,13 +28,12 @@ class CustomJpaRepositoryTest extends JpaTest {
         server = new H2();
         server.start();
         jdbcTemplate = new JdbcTemplate(server.getConnection());
-
-        entityMetaData = new EntityMetaData(Person3.class, person);
-        initForTest(entityMetaData);
     }
 
     @BeforeEach
     void setUp() {
+        entityMetaData = new EntityMetaData(Person3.class);
+        initForTest(entityMetaData);
         createTable();
     }
